pyproject.toml
[project]
name = "biological-evaluation"
version = "0.1.0"
description = "Plausability of biological scenarios through virtual patients"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "buckpass",
    "biological-scenarios-generation",
    "neo4j>=5.28.2",
    "python-libsbml>=5.20.5",
    "libroadrunner>=2.8.0",
    "openbox>=0.8.4",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.7",
    "pytest>=8.4.2",
]

[tool.uv]
override-dependencies = ["numpy==1.26.4"]

[tool.uv.sources]
buckpass = { workspace = true }
biological-scenarios-generation = { workspace = true }


[tool.uv.workspace]
members = ["packages/*"]

[tool.ruff]
line-length = 80

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D106", # Missing docstring in public nested class 
    "D107", # Missing docstring in `__init__`

    "E501", # Line too long
    "ERA001", # Found commented-out code
    "S101", # Use of `assert` detected
    "S311", # Standard pseudo-random generators are not suitable for cryptographic purposes
    "COM81", # Trailing comma missing
]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = "dynamic"
skip-magic-trailing-comma = true


[tool.basedpyright]
venvPath = "./"
venv= ".venv"
reportMissingTypeStubs = false
reportUnknownVariableType = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
extraPaths = ["./src"]
